Embed each repo on the repositories index.
The repos stored in the index attributes aren't usable as-is. I ended up
having to pull them out of `@items.find_all` while doing some digging I
discovered that while find_all is a linear time search, it does also
hint the dependency tracking system of nanoc and so it's not that
harmful to use (except in our very large case with the answers archive)

In order to get them to render correctly I also removed the default
layout from the repo item rule.